third_party/blink/web_tests/accessibility/native-slider-synthesized-events-on-action.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<style>
  .rtl {
    direction: rtl;
  }
</style>
</head>
<body>
  <input id="slider" type="range" value="0" step="1"/>

<script>
  function checkEvent(event, expected_key) {
    if (expected_key == "ArrowUp") {
      assert_equals(event.code, "ArrowUp", "event.code on " + event.type);
      assert_equals(event.key, "ArrowUp", "event.key on " + event.type);
      assert_equals(event.keyCode, 38, "event.keyCode on " + event.type);
      assert_equals(event.which, 38, "event.which on " + event.type);
    } else if (expected_key == "ArrowDown") {
      assert_equals(event.code, "ArrowDown", "event.code on " + event.type);
      assert_equals(event.key, "ArrowDown", "event.key on " + event.type);
      assert_equals(event.keyCode, 40, "event.keyCode on " + event.type);
      assert_equals(event.which, 40, "event.which on " + event.type);
    } else if (expected_key == "ArrowLeft") {
      assert_equals(event.code, "ArrowLeft", "event.code on " + event.type);
      assert_equals(event.key, "ArrowLeft", "event.key on " + event.type);
      assert_equals(event.keyCode, 37, "event.keyCode on " + event.type);
      assert_equals(event.which, 37, "event.which on " + event.type);
    } else if (expected_key == "ArrowRight") {
      assert_equals(event.code, "ArrowRight", "event.code on " + event.type);
      assert_equals(event.key, "ArrowRight", "event.key on " + event.type);
      assert_equals(event.keyCode, 39, "event.keyCode on " + event.type);
      assert_equals(event.which, 39, "event.which on " + event.type);
    }

    assert_true(event.isTrusted, "event.isTrusted on " + event.type);
    assert_equals(event.charCode, 0, "event.charCode on " + event.type);
    assert_equals(event.target, slider, "event.target on " + event.type);
    assert_equals(event.srcElement, slider, "event.srcElement on " + event.type);
    assert_true(event.bubbles, "event.bubbles on " + event.type);
    assert_false(event.defaultPrevented, "event.defaultPrevented on " + event.type);
  }

  var slider = document.getElementById("slider");
  var axSlider = accessibilityController.accessibleElementById("slider");

  test(function(t) {
    assert_true(internals.runtimeFlags.synthesizedKeyboardEventsForAccessibilityActionsEnabled);
  }, "Make sure that keyboard event synthesis is enabled");

  // Tests for left-to-right slider

  promise_test(function(t) {
    return new Promise(resolve => {
      slider.addEventListener("keydown", resolve);
      axSlider.increment();
    }).then(event => {
      checkEvent(event, "ArrowRight");
      assert_equals(slider.value, "1", "slider value after increase");
    });
  }, "check that sending an increment event to a native slider generates a right arrow keydown event");

  promise_test(function(t) {
    return new Promise(resolve => {
      slider.addEventListener("keyup", resolve);
      axSlider.increment();
    }).then(event => {
      checkEvent(event, "ArrowRight");
      assert_equals(slider.value, "2", "slider value after increase");
    });
  }, "check that sending an increment event to a native slider generates a right arrow keyup event");

  promise_test(function(t) {
    return new Promise(resolve => {
      slider.addEventListener("keydown", resolve);
      axSlider.decrement();
    }).then(event => {
      checkEvent(event, "ArrowLeft");
      assert_equals(slider.value, "1", "slider value after decrease");
    });
  }, "check that sending a decrement event to a native slider generates a left arrow keydown event");

  promise_test(function(t) {
    return new Promise(resolve => {
      slider.addEventListener("keyup", resolve);
      axSlider.decrement();
    }).then(event => {
      checkEvent(event, "ArrowLeft");
      assert_equals(slider.value, "0", "slider value after decrease");
    });
  }, "check that sending a decrement event to a native slider generates a left arrow keyup event");

  // Tests for right-to-left slider

  promise_test(function(t) {
    slider.classList.toggle("rtl");

    return new Promise(resolve => {
      slider.addEventListener("keydown", resolve);
      axSlider.increment();
    }).then(event => {
      checkEvent(event, "ArrowLeft");
      assert_equals(slider.value, "1", "slider value after increase");
    });
  }, "check that sending an increment event to a RTL native slider generates a left arrow keydown event");

  promise_test(function(t) {
    return new Promise(resolve => {
      slider.addEventListener("keyup", resolve);
      axSlider.increment();
    }).then(event => {
      checkEvent(event, "ArrowLeft");
      assert_equals(slider.value, "2", "slider value after increase");
    });
  }, "check that sending an increment event to a RTL native slider generates a left arrow keyup event");

  promise_test(function(t) {
    return new Promise(resolve => {
      slider.addEventListener("keydown", resolve);
      axSlider.decrement();
    }).then(event => {
      checkEvent(event, "ArrowRight");
      assert_equals(slider.value, "1", "slider value after decrease");
    });
  }, "check that sending a decrement event to a RTL native slider generates a right arrow keydown event");

  promise_test(function(t) {
    return new Promise(resolve => {
      slider.addEventListener("keyup", resolve);
      axSlider.decrement();
    }).then(event => {
      checkEvent(event, "ArrowRight");
      assert_equals(slider.value, "0", "slider value after decrease");
    });
  }, "check that sending a decrement event to a RTL native slider generates a right arrow keyup event");
</script>
</body>
</html>
